Remove resource server configurations, Change package name
