+++ build.gradle
@@ -15,7 +15,7 @@ apply plugin: 'eclipse'
 apply plugin: 'org.springframework.boot'
 apply plugin: 'io.spring.dependency-management'
 
-group = 'com.joonhak'
+group = 'io.joonhak'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = 11
 
@@ -31,6 +31,7 @@ dependencies {
     implementation('org.springframework.boot:spring-boot-starter-security')
     implementation('org.springframework.security:spring-security-jwt:1.0.9.RELEASE')
     implementation('org.springframework.security.oauth.boot:spring-security-oauth2-autoconfigure:2.1.1.RELEASE')
-    implementation('org.projectlombok:lombok')
+    compileOnly('org.projectlombok:lombok')
     testImplementation('org.springframework.boot:spring-boot-starter-test')
+    testCompileOnly('org.projectlombok:lombok')
 }
