--- src/main/java/com/joonhak/repository/AccountRepository.java
@@ -1,8 +1,0 @@
-package com.joonhak.repository;
-
-import com.joonhak.entity.account.Account;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccountRepository extends JpaRepository<Account, Long> {
-	Account findByUsername(String username);
-}
