+++ src/main/java/io/joonhak/OAuthServerApplication.java
@@ -1,4 +1,4 @@
-package com.joonhak;
+package io.joonhak;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
