+++ src/main/java/io/joonhak/entity/AccountDetails.java
@@ -1,12 +1,12 @@
-package com.joonhak.entity.account;
+package io.joonhak.entity;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class AccountDetails extends User {
 	
@@ -17,8 +17,8 @@ public class AccountDetails extends User {
 	}
 	
 	private static Collection<? extends GrantedAuthority> getAuthorities(List<Role> roles) {
-		var authorities = new ArrayList<GrantedAuthority>();
-		roles.forEach( r -> authorities.add(new SimpleGrantedAuthority(ROLE_PREFIX + r.getName())) );
-		return authorities;
+		return roles.stream()
+				.map( r -> new SimpleGrantedAuthority(ROLE_PREFIX + r.getName()) )
+				.collect(Collectors.toList());
 	}
 }
