+++ src/main/java/io/joonhak/repository/AccountRepository.java
@@ -1,0 +1,7 @@
+package io.joonhak.repository;
+
+import io.joonhak.entity.Account;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccountRepository extends JpaRepository<Account, String> {
+}
