+++ src/main/java/io/joonhak/repository/RoleRepository.java
@@ -1,6 +1,6 @@
-package com.joonhak.repository;
+package io.joonhak.repository;
 
-import com.joonhak.entity.account.Role;
+import io.joonhak.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<Role, Long> {
