+++ src/main/resources/application.properties
@@ -1,3 +1,2 @@
 spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
 server.port=8001
